+++ conf/script/build-system/meson/setup/compiler_version.py
@@ -1,4 +1,4 @@
-from auto_repr import auto_repr
+from auto_print import auto_repr
 from data_model import CompilerReqsSectionScheme
